new_release: Update the steps and give more detail
Instead of using Google Drive we now use HackMD and this GitHub issue
for tracking the release. Some tasks are also not very descriptive and
it also makes sense to separate the mailing list from the tweet.

Change the steps to refer to HackMD and ask for posting all links in
the GitHub issue. Also give more detail on some steps that relate to a
particular tool that is used. This helps to understand what it means
and where to search/what to ask for even if the list won't contain
comprehensive release instructions for each step.